UPDATES I changed file layout due to main.py exceeding 200 lines of code. Now api functionalities are devided by routers
